Move DeviceActivity from Arduino to JAVA-Program
Device-Activity is no longer decided on arduino side, but handled by the
Input object. Arduino gives a an array indicating inputs at wing (by
flanks) and gets an array indicating which device shall be marked as
active.
This enables use without Hardware-Controller and WriteAcces on
Device-Activity from other parts of the program.